Rendering some meters in the new layout.

+++ app/domain/device.scala
@@ -1,7 +1,7 @@
 package domain
 
 case class Register(
-    address: Short,
+    address: Int,
     minValue: Int,
     maxValue: Int) {
